MM-245 chore: todo DTO 디렉토리명 변경 (#295)
