+++ src/main/java/com/dev/moim/domain/moim/dto/todo/TodoDTO.java
@@ -1,4 +1,4 @@
-package com.dev.moim.domain.moim.dto.task;
+package com.dev.moim.domain.moim.dto.todo;
 
 import com.dev.moim.domain.moim.entity.Todo;
 import com.dev.moim.domain.moim.entity.TodoImage;
