+++ src/main/java/com/dev/moim/domain/moim/dto/todo/TodoPageDTO.java
@@ -1,4 +1,4 @@
-package com.dev.moim.domain.moim.dto.task;
+package com.dev.moim.domain.moim.dto.todo;
 
 import java.util.List;
