+++ src/main/java/com/dev/moim/domain/moim/service/impl/TodoCommandServiceImpl.java
@@ -5,7 +5,7 @@ import com.dev.moim.domain.account.entity.enums.AlarmDetailType;
 import com.dev.moim.domain.account.entity.enums.AlarmType;
 import com.dev.moim.domain.account.repository.UserRepository;
 import com.dev.moim.domain.account.service.AlarmService;
-import com.dev.moim.domain.moim.dto.task.*;
+import com.dev.moim.domain.moim.dto.todo.*;
 import com.dev.moim.domain.moim.entity.*;
 import com.dev.moim.domain.moim.entity.enums.JoinStatus;
 import com.dev.moim.domain.moim.entity.enums.TodoAssigneeStatus;
