+++ src/main/java/com/dev/moim/global/validation/validator/AddAssigneeValidator.java
@@ -1,6 +1,6 @@
 package com.dev.moim.global.validation.validator;
 
-import com.dev.moim.domain.moim.dto.task.AddTodoAssigneeDTO;
+import com.dev.moim.domain.moim.dto.todo.AddTodoAssigneeDTO;
 import com.dev.moim.domain.moim.service.MoimQueryService;
 import com.dev.moim.domain.moim.service.TodoQueryService;
 import com.dev.moim.global.validation.annotation.AddAssigneeValidation;
